Added images to multiple pages

+++ app/controllers/StationCtrl.java
@@ -15,7 +15,7 @@ public class StationCtrl extends Controller {
   }
 
   public static void addReading(Long id, int code, double temperature,
-                                double windSpeed, int windDirection, int pressure) {
+                                double windSpeed, double windDirection, double pressure) {
     Reading reading = new Reading(code, temperature, windSpeed, windDirection, pressure);
     Station station = Station.findById(id);
     station.readings.add(reading);
